Address review: emsdk CI, syscall uint64 bridge, fs/ipc fixes, MODULARIZE
- CI: use mymindstorm/setup-emsdk@v14 (4.0.1) in c-cpp and pages-wasm workflows
- Makefile: EMCC at global scope; MODULARIZE for wasm LDFLAGS
- idt_host: include idt.h
- fs.c: PATH_MAX paths, snprintf truncation check, safe realloc before capacity bump
- fl_syscall_dispatch + vm_io: uint64_t args end-to-end; rename hi in write_port_high_width
- ipc: check pthread init failures; destroy mutex on msgq early exits
- fs_facade.h: conditional pthread.h
- wasm/pre.js: globalThis.Module + DOMContentLoaded attach for output pre
- wasm/index.html: createModule(globalThis.Module)
- serve_coi: drop redundant realpath, loopback default + WASM_SERVE_BIND_ALL, recv timeouts

+++ kernel/include/fl/syscall.h
@@ -43,7 +43,7 @@ typedef enum {
 
 void fl_sys_bootstrap(void);
 void fl_sys_shutdown(void);
-long fl_syscall_dispatch(fl_syscall_no_t no, uintptr_t a0, uintptr_t a1, uintptr_t a2, uintptr_t a3);
+long fl_syscall_dispatch(fl_syscall_no_t no, uint64_t a0, uint64_t a1, uint64_t a2, uint64_t a3);
 
 #ifdef __cplusplus
 }
